Created detailed documentation using doxygen
- Updated all current comments to use java style to work with doxygen
- Added comments anywhere they were missing
- Added minor fixes / improvements as I was going through the code
- Fixed warnings

+++ PassPro/Math/VectorLibrary.h
@@ -40,7 +40,7 @@ struct Vector2
 
 	inline Vector2 Rotate(const float AngleInRadians) const { return Vector2((X * cosf(AngleInRadians)) - (Y * sinf(AngleInRadians)), (X * sinf(AngleInRadians)) + (Y * cosf(AngleInRadians))); }
 
-	inline static Vector2 GetAbs(const Vector2 Vector) { return Vector2(fabs(Vector.X), fabs(Vector.Y)); }
+	inline static Vector2 GetAbs(const Vector2 Vector) { return Vector2(fabsf(Vector.X), fabsf(Vector.Y)); }
 
 	Vector2 Normalize();
 	static Vector2 Normalize(Vector2 Vector);
